+++ file/src/primitive.rs
@@ -3,7 +3,6 @@ mod primitive_set;
 pub use primitive_set::FilePrimitiveSet;
 
 /// A primitive of a file system.
-#[repr(u8)]
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub enum Primitive {
     /// Open a file.
@@ -21,10 +20,10 @@ pub enum Primitive {
 }
 
 impl Primitive {
-    const OPEN_FILE: u8 = Self::OpenFile as _;
-    const CLOSE_FILE: u8 = Self::CloseFile as _;
-    const READ_FILE: u8 = Self::ReadFile as _;
-    const WRITE_FILE: u8 = Self::WriteFile as _;
-    const DELETE_FILE: u8 = Self::DeleteFile as _;
-    const EXISTS_FILE: u8 = Self::ExistsFile as _;
+    const OPEN_FILE: usize = Self::OpenFile as _;
+    const CLOSE_FILE: usize = Self::CloseFile as _;
+    const READ_FILE: usize = Self::ReadFile as _;
+    const WRITE_FILE: usize = Self::WriteFile as _;
+    const DELETE_FILE: usize = Self::DeleteFile as _;
+    const EXISTS_FILE: usize = Self::ExistsFile as _;
 }
