+++ file/src/primitive/primitive_set.rs
@@ -1,8 +1,7 @@
 use super::Primitive;
 use crate::FileSystem;
 use heapless::Vec;
-use stak_vm::Error;
-use stak_vm::{Memory, Number, PrimitiveSet, Value};
+use stak_vm::{Error, Memory, Number, PrimitiveSet, Value};
 
 const PATH_SIZE: usize = 64;
 
@@ -53,7 +52,7 @@ impl<T: FileSystem> FilePrimitiveSet<T> {
 impl<T: FileSystem> PrimitiveSet for FilePrimitiveSet<T> {
     type Error = Error;
 
-    fn operate(&mut self, memory: &mut Memory, primitive: u8) -> Result<(), Self::Error> {
+    fn operate(&mut self, memory: &mut Memory, primitive: usize) -> Result<(), Self::Error> {
         match primitive {
             Primitive::OPEN_FILE => Self::operate_option(memory, |memory| {
                 let [list, output] = memory.pop_many();
