+++ process_context/src/primitive.rs
@@ -3,7 +3,6 @@ mod primitive_set;
 pub use primitive_set::ProcessContextPrimitiveSet;
 
 /// A primitive of process context.
-#[repr(u8)]
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub enum Primitive {
     /// A command line.
@@ -13,6 +12,6 @@ pub enum Primitive {
 }
 
 impl Primitive {
-    const COMMAND_LINE: u8 = Self::CommandLine as _;
-    const ENVIRONMENT_VARIABLES: u8 = Self::EnvironmentVariables as _;
+    const COMMAND_LINE: usize = Self::CommandLine as _;
+    const ENVIRONMENT_VARIABLES: usize = Self::EnvironmentVariables as _;
 }
